refactor/moved logic to custom hook for larger components
